refactor: public path runtime module
